Make utils module
