+++ src/sknnr_spatial/image/_base.py
@@ -7,7 +7,6 @@ from types import ModuleType
 from typing import TYPE_CHECKING
 
 import numpy as np
-import xarray as xr
 from numpy.typing import NDArray
 
 if TYPE_CHECKING:
@@ -185,15 +184,3 @@ def kneighbors(
 ) -> None:
     msg = f"kneighbors is not implemented for type `{X_image.__class__.__name__}`."
     raise NotImplementedError(msg)
-
-
-def is_image_type(X: ImageType) -> bool:
-    # Feature array images must have exactly 3 dimensions: (y, x, band) or (band, y, x)
-    if isinstance(X, (np.ndarray, xr.DataArray)):
-        return X.ndim == 3
-
-    # Feature Dataset images must have exactly 2 dimensions: (x, y)
-    if isinstance(X, xr.Dataset):
-        return len(X.dims) == 2
-
-    return False
