+++ src/sknnr_spatial/utils/image.py
@@ -1,0 +1,16 @@
+import numpy as np
+import xarray as xr
+
+from ..types import ImageType
+
+
+def is_image_type(X: ImageType) -> bool:
+    # Feature array images must have exactly 3 dimensions: (y, x, band) or (band, y, x)
+    if isinstance(X, (np.ndarray, xr.DataArray)):
+        return X.ndim == 3
+
+    # Feature Dataset images must have exactly 2 dimensions: (x, y)
+    if isinstance(X, xr.Dataset):
+        return len(X.dims) == 2
+
+    return False
